src/main/java/util/common/UrlCoderUtils.java
package util.common;

import util.helper.Charset;

import java.io.UnsupportedEncodingException;
import java.net.URLDecoder;
import java.net.URLEncoder;

/**
 * Created by liangjj on 2017/9/20.
 */
class UrlCoderUtils {
    private UrlCoderUtils() {
    }

    /**
     * URL 编码
     *
     * @param s
     * @return
     * @throws UnsupportedEncodingException
     */
    static String urlEncoder(String s) throws UnsupportedEncodingException {
        return URLEncoder.encode(s, Charset.UTF8.toString());
    }

    /**
     * URL 解码
     *
     * @param s
     * @return
     * @throws UnsupportedEncodingException
     */
    static String urlDecoder(String s) throws UnsupportedEncodingException {
        return URLDecoder.decode(s, Charset.UTF8.toString());
    }
}
